Merged in Changeset 1 from sibling project.
WinAPI code added to this revision is identical with that from the other
repository. The rest of the code has minor changes.
Original timestamp:                                      8/ 4/2011  5:49:26pm

Original commit message:
"now enumerates printers and their OpenGL contexts.
OpenGL contexts seem to be independent of the device."

Is it possible that ATI's OpenGL implementation supports memory and printer
contexts, but nVidia's doesn't?

--HG--
branch : WinAPI version